+++ Changelog.md
@@ -21,6 +21,7 @@
   `People#get_reassign_progress`,
   `People#terminate_reassign`
 * Add `Community#delete_event`
+* Add `Community#delete_thread`
 * Remove usage of `httmultiparty`
 
 ### Fixes